+++ .idea/dataSources/4829c2d2-745e-464f-89f4-44172b45effe/storage_v2/_src_/schema/sys.zb4BAA.meta
@@ -1,0 +1,2 @@
+#n:sys
+!<md> [null, 0, null, null, -2147483648, -2147483648]
