+++ projectboard2/src/main/java/com/koreait/projectboard2/Projectboard2Application.java
@@ -1,0 +1,13 @@
+package com.koreait.projectboard2;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Projectboard2Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Projectboard2Application.class, args);
+	}
+
+}
