+++ projectboard2/src/test/java/com/koreait/projectboard2/Projectboard2ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.koreait.projectboard2;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Projectboard2ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
